Code example in the book
It is a sum up of all the codes example in one .py file. 
The code is not written in the OOP style, but in basic function only. 
As the during the step by step illustration of Python implementation, i found out that it is not practical to write in OOP since i need to break down the code structure to explain while having OOP design pattern could lead to a misunderstanding of the purpose / variables, parameters, or even introduce confusion.
There will be a separated code base written in OOP style, more like a MC simulation engine prototype in itself.